+++ src/utils/logger.cpp
@@ -5,6 +5,7 @@
 #include "utils/logger.h"
 #include "utils/limits.h" // NOLINT(misc-include-cleaner)
 #include <ctime>
+#include <cstring>
 
 #ifdef _WIN32
 #include <cassert>
@@ -20,6 +21,7 @@ namespace dncdbg
 {
 
 std::mutex Logger::m_logMutex;
+LogLevel Logger::m_logLevel = LogLevel::INF;
 
 namespace
 {
@@ -83,6 +85,36 @@ void Logger::OpenLogStream(const char *fileName)
     GetLogStream().open(fileName, std::ios::app);
 }
 
+void Logger::SetLogLevel(const char *level)
+{
+    if (level == nullptr)
+    {
+        return;
+    }
+
+    if (strcmp(level, "0") == 0 || strcmp(level, "DEBUG") == 0)
+    {
+        m_logLevel = LogLevel::DBG;
+    }
+    else if (strcmp(level, "1") == 0 || strcmp(level, "INFO") == 0)
+    {
+        m_logLevel = LogLevel::INF;
+    }
+    else if (strcmp(level, "2") == 0 || strcmp(level, "WARNING") == 0)
+    {
+        m_logLevel = LogLevel::WRN;
+    }
+    else if (strcmp(level, "3") == 0 || strcmp(level, "ERROR") == 0)
+    {
+        m_logLevel = LogLevel::ERR;
+    }
+    else
+    {
+        m_logLevel = LogLevel::INF;
+        LOGW(log << "Unknown log level '" << level << "', using default INFO level");
+    }
+}
+
 // Function should form output line like this:
 //
 // 1500636976.777 I(P 2293, T 2293): udev.c:64 uevent_control_cb() > Set udev monitor buffer size 131072
@@ -91,33 +123,29 @@ void Logger::OpenLogStream(const char *fileName)
 // |              ` log level         ` file name          ` function name
 // `--- time sec.msec
 //
-void Logger::LogPrint(LogPriority prio, const char *file, int line, const char *func, const LoggerCallback &cb)
+void Logger::LogPrint(LogLevel level, const char *file, int line, const char *func, const LoggerCallback &cb)
 {
+    if (level < m_logLevel ||
+        !GetLogStream().good())
+    {
+        return;
+    }
+
     struct timespec ts{};
     clock_gettime(CLOCK_MONOTONIC, &ts);
 
     const std::scoped_lock<std::mutex> lock(m_logMutex);
 
-    if (prio == LogPriority::DEF)
+    char levelSymb = 'I';
+    if (level >= LogLevel::DBG && level <= LogLevel::ERR)
     {
-        prio = LogPriority::INF;
-    }
-
-    char level = 'I';
-    if (prio >= LogPriority::DBG && prio <= LogPriority::FTL)
-    {
-        level = "DIWEF"[static_cast<uint8_t>(prio) - static_cast<uint8_t>(LogPriority::DBG)];
-    }
-
-    if (!GetLogStream().good())
-    {
-        return;
+        levelSymb = "DIWE"[static_cast<uint8_t>(level)];
     }
 
     GetLogStream() << (ts.tv_sec & MAX_TIMESTAMP_SECONDS) << '.'
                    << std::dec << std::setfill('0') << std::setw(3)
                    << (ts.tv_nsec / NSEC_TO_MSEC) << ' '
-                   << level << "(P" << std::setw(4) << get_pid()
+                   << levelSymb << "(P" << std::setw(4) << get_pid()
                    << ", T" << std::setw(4) << get_tid() << "): "
                    << file << ":" << line << " " << func << "() > ";